requirements.txt
aiohttp==3.8.4 ; python_version >= "3.8" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.8" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.8" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4"
charset-normalizer==3.1.0 ; python_version >= "3.8" and python_version < "4.0"
click==7.1.2 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
datasets==2.11.0 ; python_version >= "3.8" and python_version < "4.0"
dill==0.3.6 ; python_version >= "3.8" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.8" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.8" and python_version < "4.0"
fastapi==0.68.2 ; python_version >= "3.8" and python_version < "4.0"
filelock==3.12.0 ; python_version >= "3.8" and python_version < "4.0"
fire==0.5.0 ; python_version >= "3.8" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.8" and python_version < "4.0"
fsspec[http]==2023.4.0 ; python_version >= "3.8" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.8" and python_version < "4.0"
h11==0.9.0 ; python_version >= "3.8" and python_version < "4.0"
httptools==0.1.2 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.8" and python_version < "4.0"
huggingface-hub==0.13.4 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "4.0"
lxml==4.9.2 ; python_version >= "3.8" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
multiprocess==0.70.14 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.2 ; python_version >= "3.8" and python_version < "4.0"
openai==0.27.4 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.8" and python_version < "4.0"
pandas==2.0.0 ; python_version >= "3.8" and python_version < "4.0"
pyarrow==11.0.0 ; python_version >= "3.8" and python_version < "4.0"
pydantic==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
pydantic[email]==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
pyfunctional==1.4.3 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.8" and python_version < "4.0"
python-slugify==8.0.1 ; python_version >= "3.8" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.8" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.8" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.8" and python_version < "4"
responses==0.18.0 ; python_version >= "3.8" and python_version < "4.0"
setuptools==67.6.1 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.8" and python_version < "4.0"
starlette==0.14.2 ; python_version >= "3.8" and python_version < "4.0"
tabulate==0.9.0 ; python_version >= "3.8" and python_version < "4.0"
termcolor==2.2.0 ; python_version >= "3.8" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.8" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.8" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.8" and python_version < "4"
uvicorn==0.11.8 ; python_version >= "3.8" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.8" and python_version < "4.0"
websockets==8.1 ; python_version >= "3.8" and python_version < "4.0"
xxhash==3.2.0 ; python_version >= "3.8" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
